refactor(DiscountAmount, MinimumOrderAmount): 금액 타입을 Long으로 일치

+++ src/main/java/coupon/coupon/domain/Coupon.java
@@ -50,7 +50,7 @@ public class Coupon extends BaseTimeEntity {
     public Coupon(
             final String name,
             final Long discountAmount,
-            final Integer minimumOrderAmount,
+            final Long minimumOrderAmount,
             final LocalDate issueStartDate,
             final LocalDate issueEndDate
     ) {
